Static provider
Static provider serves for mixing already provisioned hosts (pets)
with dynamically provisioned hosts. Alternatively all hosts can
be static. This can be useful for generating the outputs and using
other features.

Almost all operations are fake given that the hosts exist and cannot
be deleted by static provider. The most important parts is to provide
a name and ip.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>