Add basic image validation to AWS and Beaker
Added basic image validation to AWS and Beaker provider
When os was not known in provisioning config, null/None
was taken and set as required image even for Openstack.

Value of 'os' is passed to req when image is not in place.
Then the 'os' value is seen in ValidationError messages
even instead of None/null in Openstack, AWS and Beaker.

Support image/distro definition from metadata.
And flag it with meta_distro/meta_image when used.

For AWS check image availability with boto.
When image is not present raise ValidationError

Signed-off-by: Tibor Dudlák <tdudlak@redhat.com>