Fix OpenStack provisioning status translation
It is not needed to use a STATUS_MAP for translation status in
prov_result_to_host_data as it is done also in to_host of parent
class. Double translation doesn't work and thus everything provisioned
had incorrecly status : "other".

This has then a negative effect on ssh actione where no host is reported
as active.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>