feat: Support size definition in metadata
Enable flavor translation based on size defined
in the metadata file. If the metadata does not
contain the size key for host entry flavor
will be based on host group.

This enables possibility to provision bigger/smaller
instances for specific usecases with keeping
group in place for each host so 'roles' of VMs
in the automation remain as expected based on groups.

Signed-off-by: Tibor Dudlák <tdudlak@redhat.com>