feat: Enable config drive in OpenStack
By default Nova's metadata informatation if retrieved via HTTP, setting
`enable_config_drive` to True in provisioning-config will make metadata
available via a config drive, a special drive that are attached to an
instance when it boots.

This is helpful when connection is not stable during host booting stage,
host would be unreachable because cloud-init is failing to injecting
SSH keys into it.

This reduces the chance of hosts being provisioned but not accessible.

Signed-off-by: Armando Neto <abiagion@redhat.com>