feat: check ssh connectivity outside of pares_error_hosts more than once
Move the ssh connectivity check outside of the
pares_error_hosts method to distinguish the functionality
and to be able to gather asynchronous results later.

Retry the ssh connection after if fails to not destroy
slower systems which may not have set up the proper
ssh key into allowed keys.

Signed-off-by: Tibor Dudlák <tdudlak@redhat.com>