feat: Use GlobalContext class in runtime
mrack now uses the GlobalContext object to store
the information for every particular run.
This object is defined in the context.py
And should be initialized for the run.

BREAKING CHANGE: Using GlobalContext class in runtime
allows us to use mrack actions without parameters
Up/Destroy/... which simplify the workflow in scripts
so mrack can be used as library in python automation.

Signed-off-by: Tibor Dudlák <tdudlak@redhat.com>