feat(Virt): use generated run IDs for parallel jobs
It may be common use case to run on one host machine two mrack jobs
with job metadata files which share hostnames. This may result in
conflict in testcloud as it identifies the vms by provided name. So
this patch is generating a pseudo unique run ID which is used as a
prefix for testcloud vm names and thus making it whole unique and
conflict resilient.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>