fix: do not end provisioning if there are no resources
Providers ended provisioning early by throwing an exception
ValidationError which caused mrack to end provisioning
even when retry strategy has been enabled.
Now we create error Host object to continue provisioning with.
Also fixed condition when we need to cleanup the
error hosts for the next retry of the provisioning.

Signed-off-by: Tibor Dudlák <tdudlak@redhat.com>