feat(aws): defining AMIs by tags
Add a new way to define an AWS EC2 image in a provisiong config in
addition to ami ID.

The following is valid:

```
images:
  fedora-34: ami-0d3c5199abb29a7ae
  fedora-35: { tag: { name: compose, value: mrack-fedora-35-latest } }
```

This allows for a seperate tool to upload new image with the same
purpose and mrack automatically picks the latest one with the given
tag.

Currently only one tag is supported and it search in all images.

Possible improvements:
- more tags
- limiting to private images

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>