feat: copyign meta_ attributes from host to ansible inventory
Make it possible to define arbitratry attributes prefix with meta_
in host section of job metadata file.

All of these attributes are then copied to host part in ansible
inventory output.

It can serve for creation of playbooks which consumes these attributes.

Addtional, but probably rare use case, is to override the meta_ attrs
which are derived from other others by mrack.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>