fix(pytest-multihost): crash when group is not defined
It should possible to define host groups either via "group" or "groups"
variable. But pytest-multihost output was crashing when it was done via
"groups".

This fix makes it not to crash if both are missing. It also makes it
possible to use first group from "groups" if groups are defined -
similar as in Ansible inventory output.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>