feat(aws): Add owner/lifetime info in VM's metadata
Owner and lifetime info will be added into VM's metadata
for AWS provider. This info will give info who
owns the VM and what is the lifetime of the VM which will
be used for notifying owners and cleaning up the resources
based on lifetime of the VMs.
If owner info not fetched, user notified with custom error
message for follow up action to take up.

Signed-off-by: ksiddiqu <ksiddiqu@redhat.com>