chore(release): Add PyPI action & extract copr step
Due to latest changes to python semantic release
PyPI release is no longer supported and the separate
action (https://github.com/pypa/gh-action-pypi-publish)
is recomended to use while releasing to PyPI.
For the trusted publishing we had to set up the PyPI
account owning mrack to trust neoave/mrack repository
which I did and set up the actions jobs to build
the python package and trigger a truster build.
As an addition i have extracted copr to separate job.

Some resources:
https://docs.pypi.org/trusted-publishers/
https://python-semantic-release.readthedocs.io/en/latest/index.html
https://github.com/pypa/gh-action-pypi-publish
https://python-semantic-release.readthedocs.io/en/latest/migrating_from_v7.html

we had to move from v7 to 8 to use PyPI and have
build steps separate as v7 does that inplace.

Signed-off-by: Tibor Dudlák <tibor.dudlak@gmail.com>