fix(beaker): supress 10_avc_check restraint plugin
It has happpened that this plugin sometimes ran after the dummy task, it reported fail
that there was some AVC (probabaly from other thing) which then failed the job and thus
mrack treated this as a provisioning failure.

This patch instructs restraint to not run this plugin and thus avoid this situation.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>